% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skillhelpers.R
\name{collapse_states}
\alias{collapse_states}
\title{Collapse states with items with numeric prefixes}
\usage{
collapse_states(bloated_states)
}
\arguments{
\item{bloated_states}{a matrix representing the knowledge structure with numeric prefixes in the item names}
}
\value{
a matrix representing the knowledge structure with collapsed states
}
\description{
Combines items that differ only in numeric prefix into single items, and then reasseses the states
}
