
\name{conjunctive.as.disjunctive}
\alias{conjunctive.as.disjunctive}
\title{Convert a conjunctive skill function to a disjunctive one}
\usage{
conjunctive.as.disjunctive(sf)
}
\arguments{
\item{sf}{a data frame representing a conjunctive skill function}
}
\value{
a data frame representing the equivalent disjunctive skill function
}
\description{
Convert a conjunctive skill function to a disjunctive one
}
