
\name{disjunctive.as.conjunctive}
\alias{disjunctive.as.conjunctive}
\title{Convert a disjunctive skill function to a conjunctive one}
\usage{
disjunctive.as.conjunctive(sf)
}
\arguments{
\item{sf}{a data frame representing a disjunctive skill function}
}
\value{
a data frame representing the equivalent conjunctive skill function
}
\description{
Convert a disjunctive skill function to a conjunctive one
}
