
\name{isCon}
\alias{isCon}
\title{Check if a skill function is conjunctive}
\usage{
isCon(skillfun, itemID)
}
\arguments{
\item{skillfun}{a data frame representing a skill function}

\item{itemID}{index of the column in \code{skillfun} that holds the item indicator}
}
\value{
logical indicating whether the skill function is conjunctive
}
\description{
Check if a skill function is conjunctive
}
