
\name{isDis}
\alias{isDis}
\title{Check if a skill function is disjunctive}
\usage{
isDis(skillfun, itemID)
}
\arguments{
\item{skillfun}{a data frame representing a skill function}

\item{itemID}{index of the column in \code{skillfun} that holds the item indicator}
}
\value{
logical indicating whether the skill function is disjunctive
}
\description{
Check if a skill function is disjunctive
}
