
\name{K.are.same}
\alias{K.are.same}
\title{Compare two knowledge structures for equality}
\usage{
K.are.same(K1, K2)
}
\arguments{
\item{K1}{first knowledge structure}

\item{K2}{second knowledge structure}
}
\value{
logical indicating whether the knowledge structures are equal. 
}
\description{
Checks if two knowledge structures contain the same states.
}
