% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blim.R
\name{plot.blim}
\alias{plot.blim}
\title{Diagnostic Plot for Basic Local Independence Models}
\usage{
\method{plot}{blim}(
  x,
  xlab = "Predicted response probabilities",
  ylab = "Deviance residuals",
  ...
)
}
\arguments{
\item{x}{an object of class \code{blim}, typically the result of a call to
\code{\link{blim}}.}

\item{xlab, ylab, }{graphical parameters passed to plot.}

\item{list()}{graphical parameters passed to plot.}
}
\description{
Plots BLIM residuals against fitted values.
}
\details{
The deviance residuals are plotted against the predicted response
probabilities for each response pattern.
}
\examples{

## Compare MD and MDML estimation

data(DoignonFalmagne7)
blim1 <- blim(DoignonFalmagne7$K, DoignonFalmagne7$N.R, method="MD")
blim2 <- blim(DoignonFalmagne7$K, DoignonFalmagne7$N.R, method="MDML")

par(mfrow = 1:2)      # residuals versus fitted values
plot(blim1, main = "MD estimation",   ylim = c(-4, 4))
plot(blim2, main = "MDML estimation", ylim = c(-4, 4))

}
\seealso{
\code{\link{blim}}, \code{\link{residuals.blim}}.
}
\keyword{models}
