
\name{make_con_to_dis}
\alias{make_con_to_dis}
\title{Convert a conjunctive skill function to a disjunctive one}
\usage{
make_con_to_dis(sf)
}
\arguments{
\item{sf}{a data frame representing a conjunctive skill function}
}
\value{
a data frame representing the equivalent disjunctive skill function
}
\description{
Convert a conjunctive skill function to a disjunctive one
}
