
\name{make_dis_to_con}
\alias{make_dis_to_con}
\title{Convert a disjunctive skill function to a conjunctive one}
\usage{
make_dis_to_con(sf)
}
\arguments{
\item{sf}{a data frame representing a disjunctive skill function}
}
\value{
a data frame representing the equivalent conjunctive skill function
}
\description{
Convert a disjunctive skill function to a conjunctive one
}
